refactor(broker): extract shared SCM_RIGHTS consumer-client helpers
Factor `send_message_with_fd` / `recv_message_with_fd` / `send_request`
/ `connect_to_broker` out of their three homes (streamlib-broker,
streamlib-python-native, streamlib-deno-native) into a new
`streamlib-broker-client` crate with `libc` + `serde_json` as its
only deps. The broker server re-exports the helpers so existing call
sites keep working; the two polyglot cdylibs drop their `mod wire`
copies and alias the shared crate.

Closes #422
