+++ Cargo.toml
@@ -4,6 +4,7 @@ members = [
     "xtask",                    # Development tasks (cargo xtask <command>)
     "libs/streamlib",           # All-in-one: core, apple (feature-gated)
     "libs/streamlib-broker",    # Broker service for cross-process coordination (macOS)
+    "libs/streamlib-broker-client", # Consumer-side wire helpers for the broker's Unix-socket + SCM_RIGHTS protocol (Linux)
     "libs/streamlib-cli",       # Command-line interface for StreamLib runtime
     "libs/streamlib-runtime",   # Standalone runtime binary (spawned by CLI)
     "libs/streamlib-codegen-shared", # Execution types shared between streamlib and streamlib-macros
