+++ libs/streamlib-broker-client/Cargo.toml
@@ -1,0 +1,19 @@
+# Copyright (c) 2025 Jonathan Fontanez
+# SPDX-License-Identifier: BUSL-1.1
+
+[package]
+name = "streamlib-broker-client"
+description = "Consumer-side wire helpers for the StreamLib broker's Unix-socket + SCM_RIGHTS protocol"
+version.workspace = true
+edition.workspace = true
+authors.workspace = true
+license-file.workspace = true
+repository.workspace = true
+
+[lib]
+name = "streamlib_broker_client"
+path = "src/lib.rs"
+
+[target.'cfg(target_os = "linux")'.dependencies]
+libc.workspace = true
+serde_json.workspace = true
