+++ libs/streamlib-python-native/Cargo.toml
@@ -23,6 +23,7 @@ serde_json.workspace = true
 
 [target.'cfg(target_os = "linux")'.dependencies]
 libc.workspace = true
+streamlib-broker-client = { path = "../streamlib-broker-client" }
 
 [dev-dependencies]
 # Broker unit tests spin up a real UnixSocketSurfaceService + client in-process.
