+++ libs/streamlib-python-native/src/lib.rs
@@ -1590,200 +1590,11 @@ mod broker_client {
     }
 
     // =========================================================================
-    // Wire helpers — duplicated from streamlib_broker::unix_socket_service to
-    // keep this cdylib's dependency surface minimal. Kept byte-compatible by
-    // design (see unit tests in both crates).
+    // Wire helpers come from the shared `streamlib-broker-client` crate so the
+    // broker server and every polyglot cdylib speak a single-sourced protocol.
+    // Aliased as `wire` here to preserve the original call-site shape.
     // =========================================================================
-    mod wire {
-        use std::os::unix::io::{AsRawFd, RawFd};
-        use std::os::unix::net::UnixStream;
-
-        pub fn send_message_with_fd(
-            stream: &UnixStream,
-            data: &[u8],
-            fd: Option<RawFd>,
-        ) -> std::io::Result<()> {
-            // Length prefix (4 bytes big-endian)
-            let len_bytes = (data.len() as u32).to_be_bytes();
-            let mut len_iov = libc::iovec {
-                iov_base: len_bytes.as_ptr() as *mut libc::c_void,
-                iov_len: 4,
-            };
-            let mut len_msg: libc::msghdr = unsafe { std::mem::zeroed() };
-            len_msg.msg_iov = &mut len_iov;
-            len_msg.msg_iovlen = 1;
-            let n = unsafe { libc::sendmsg(stream.as_raw_fd(), &len_msg, 0) };
-            if n < 0 {
-                return Err(std::io::Error::last_os_error());
-            }
-
-            // Payload + optional SCM_RIGHTS
-            let mut iov = libc::iovec {
-                iov_base: data.as_ptr() as *mut libc::c_void,
-                iov_len: data.len(),
-            };
-            let mut msg: libc::msghdr = unsafe { std::mem::zeroed() };
-            msg.msg_iov = &mut iov;
-            msg.msg_iovlen = 1;
-
-            const CMSG_SPACE_SIZE: usize =
-                unsafe { libc::CMSG_SPACE(std::mem::size_of::<RawFd>() as u32) } as usize;
-            #[repr(C)]
-            union CmsgBuf {
-                buf: [u8; CMSG_SPACE_SIZE],
-                _align: libc::cmsghdr,
-            }
-            let mut cmsg_buf = CmsgBuf {
-                buf: [0u8; CMSG_SPACE_SIZE],
-            };
-
-            if let Some(send_fd) = fd {
-                msg.msg_control = unsafe { cmsg_buf.buf.as_mut_ptr() } as *mut libc::c_void;
-                msg.msg_controllen = CMSG_SPACE_SIZE;
-                let cmsg_ptr = unsafe { libc::CMSG_FIRSTHDR(&msg) };
-                if !cmsg_ptr.is_null() {
-                    unsafe {
-                        (*cmsg_ptr).cmsg_level = libc::SOL_SOCKET;
-                        (*cmsg_ptr).cmsg_type = libc::SCM_RIGHTS;
-                        (*cmsg_ptr).cmsg_len =
-                            libc::CMSG_LEN(std::mem::size_of::<RawFd>() as u32) as usize;
-                        let fd_ptr = libc::CMSG_DATA(cmsg_ptr) as *mut RawFd;
-                        *fd_ptr = send_fd;
-                    }
-                    msg.msg_controllen = CMSG_SPACE_SIZE;
-                }
-            }
-
-            let n = unsafe { libc::sendmsg(stream.as_raw_fd(), &msg, 0) };
-            if n < 0 {
-                return Err(std::io::Error::last_os_error());
-            }
-            Ok(())
-        }
-
-        pub fn recv_message_with_fd(
-            stream: &UnixStream,
-            msg_len: usize,
-        ) -> std::io::Result<(Vec<u8>, Option<RawFd>)> {
-            const CMSG_SPACE_SIZE: usize =
-                unsafe { libc::CMSG_SPACE(std::mem::size_of::<RawFd>() as u32) } as usize;
-            #[repr(C)]
-            union CmsgBuf {
-                buf: [u8; CMSG_SPACE_SIZE],
-                _align: libc::cmsghdr,
-            }
-            let mut cmsg_buf = CmsgBuf {
-                buf: [0u8; CMSG_SPACE_SIZE],
-            };
-
-            let mut buf = vec![0u8; msg_len];
-            let mut iov = libc::iovec {
-                iov_base: buf.as_mut_ptr() as *mut libc::c_void,
-                iov_len: msg_len,
-            };
-            let mut msg: libc::msghdr = unsafe { std::mem::zeroed() };
-            msg.msg_iov = &mut iov;
-            msg.msg_iovlen = 1;
-            msg.msg_control = unsafe { cmsg_buf.buf.as_mut_ptr() } as *mut libc::c_void;
-            msg.msg_controllen = CMSG_SPACE_SIZE;
-
-            let n = unsafe { libc::recvmsg(stream.as_raw_fd(), &mut msg, 0) };
-            if n < 0 {
-                return Err(std::io::Error::last_os_error());
-            }
-            if n == 0 {
-                return Err(std::io::Error::new(
-                    std::io::ErrorKind::UnexpectedEof,
-                    "Connection closed",
-                ));
-            }
-            if msg.msg_flags & libc::MSG_CTRUNC != 0 {
-                return Err(std::io::Error::new(
-                    std::io::ErrorKind::InvalidData,
-                    "SCM_RIGHTS control message truncated",
-                ));
-            }
-
-            let mut total_read = n as usize;
-            while total_read < msg_len {
-                let remaining = &mut buf[total_read..];
-                let n = unsafe {
-                    libc::read(
-                        stream.as_raw_fd(),
-                        remaining.as_mut_ptr() as *mut libc::c_void,
-                        remaining.len(),
-                    )
-                };
-                if n <= 0 {
-                    return Err(std::io::Error::new(
-                        std::io::ErrorKind::UnexpectedEof,
-                        "Connection closed during message read",
-                    ));
-                }
-                total_read += n as usize;
-            }
-
-            let mut received_fd = None;
-            let mut cmsg_ptr = unsafe { libc::CMSG_FIRSTHDR(&msg) };
-            while !cmsg_ptr.is_null() {
-                let cmsg = unsafe { &*cmsg_ptr };
-                if cmsg.cmsg_level == libc::SOL_SOCKET && cmsg.cmsg_type == libc::SCM_RIGHTS {
-                    let fd_ptr = unsafe { libc::CMSG_DATA(cmsg_ptr) } as *const RawFd;
-                    received_fd = Some(unsafe { *fd_ptr });
-                }
-                cmsg_ptr = unsafe { libc::CMSG_NXTHDR(&msg, cmsg_ptr) };
-            }
-            // Use `buf` now to keep the underlying allocation alive.
-            let _ = &buf;
-            let mut read_buf = vec![0u8; msg_len];
-            read_buf.copy_from_slice(&buf[..msg_len]);
-            Ok((read_buf, received_fd))
-        }
-
-        pub fn send_request(
-            stream: &UnixStream,
-            request: &serde_json::Value,
-            fd: Option<RawFd>,
-        ) -> std::io::Result<(serde_json::Value, Option<RawFd>)> {
-            let request_bytes = serde_json::to_vec(request).map_err(|e| {
-                std::io::Error::new(
-                    std::io::ErrorKind::Other,
-                    format!("Failed to serialize request: {}", e),
-                )
-            })?;
-            send_message_with_fd(stream, &request_bytes, fd)?;
-
-            // Read response length prefix (4 bytes big-endian)
-            let mut len_buf = [0u8; 4];
-            let mut total = 0;
-            while total < 4 {
-                let n = unsafe {
-                    libc::read(
-                        stream.as_raw_fd(),
-                        len_buf[total..].as_mut_ptr() as *mut libc::c_void,
-                        4 - total,
-                    )
-                };
-                if n <= 0 {
-                    return Err(std::io::Error::new(
-                        std::io::ErrorKind::UnexpectedEof,
-                        "Failed to read response length",
-                    ));
-                }
-                total += n as usize;
-            }
-            let response_len = u32::from_be_bytes(len_buf) as usize;
-            let (response_bytes, response_fd) = recv_message_with_fd(stream, response_len)?;
-            let response: serde_json::Value = serde_json::from_slice(&response_bytes)
-                .map_err(|e| {
-                    std::io::Error::new(
-                        std::io::ErrorKind::InvalidData,
-                        format!("Invalid JSON response: {}", e),
-                    )
-                })?;
-            Ok((response, response_fd))
-        }
-    }
+    use streamlib_broker_client as wire;
 
     fn c_str_to_string(ptr: *const c_char) -> Option<String> {
         if ptr.is_null() {
